backend_sample/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
langgraph>=1.0,<2.0
langchain>=1.0,<2.0
langchain-core>=1.0,<2.0
langchain-openai>=1.0,<2.0
